+++ minos/common/__init__.py
@@ -6,7 +6,7 @@ This file is part of minos framework.
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
 
-__version__ = "0.0.9"
+__version__ = "0.0.11"
 
 from .broker import (
     MinosBaseBroker,
@@ -63,6 +63,11 @@ from .messages import (
     MinosRPCHeadersRequest,
     MinosRPCResponse,
 )
+from .meta import (
+    classproperty,
+    property_or_classproperty,
+    self_or_classmethod,
+)
 from .model import (
     ARRAY,
     BOOLEAN,
@@ -103,6 +108,7 @@ from .protocol import (
     MinosAvroProtocol,
     MinosAvroValuesDatabase,
     MinosBinaryProtocol,
+    MinosJsonBinaryProtocol,
 )
 from .repository import (
     MinosInMemoryRepository,