+++ minos/common/model/command.py
@@ -6,10 +6,9 @@ This file is part of minos framework.
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
 
-from minos.common.model import (
-    Aggregate,
+from .abc import (
+    MinosModel,
 )
-
 from .command_reply import (
     CommandReply,
 )
@@ -20,5 +19,5 @@ class Command(CommandReply):
 
     reply_on: str
 
-    def __init__(self, topic: str, items: list[Aggregate], saga_id: str, task_id: str, reply_on: str, *args, **kwargs):
+    def __init__(self, topic: str, items: list[MinosModel], saga_id: str, task_id: str, reply_on: str, *args, **kwargs):
         super().__init__(topic, items, *args, saga_id=saga_id, task_id=task_id, reply_on=reply_on, **kwargs)