+++ minos/common/model/command_reply.py
@@ -6,10 +6,9 @@ This file is part of minos framework.
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
 
-from minos.common.model import (
-    Aggregate,
+from .abc import (
+    MinosModel,
 )
-
 from .event import (
     Event,
 )
@@ -21,5 +20,5 @@ class CommandReply(Event):
     saga_id: str
     task_id: str
 
-    def __init__(self, topic: str, items: list[Aggregate], saga_id: str, task_id: str, *args, **kwargs):
+    def __init__(self, topic: str, items: list[MinosModel], saga_id: str, task_id: str, *args, **kwargs):
         super().__init__(topic, items, *args, saga_id=saga_id, task_id=task_id, **kwargs)