+++ minos/common/protocol/__init__.py
@@ -6,10 +6,13 @@ This file is part of minos framework.
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
 
-from .abstract import (
+from .abc import (
     MinosBinaryProtocol,
 )
 from .avro import (
     MinosAvroProtocol,
     MinosAvroValuesDatabase,
+)
+from .json import (
+    MinosJsonBinaryProtocol,
 )